requirements.txt
altgraph==0.17.4
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
packaging==23.2
pefile==2023.2.7
pillow==10.2.0
pycparser==2.21
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PyMuPDFb==1.23.9
PySocks==1.7.1
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
reportlab==4.0.9
requests==2.31.0
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.0
webdriver-manager==4.0.1
wsproto==1.2.0
